Renaming "director" to "screening", remove prod
The screening directory will host the "http://localhost:8081/screening/..." files so screening makes more sense. I know we now have screening/server/screening, but I think this is less confusing than director.

Also, removed the production mode for the server. We don't need it. No one is using it and it is probably broken.